pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pySerialMux"
version = "0.1.0"
description = "Drop-in 1-to-N multiplexer for pySerial hardware serial interfaces"
requires-python = ">=3.8"
dependencies = ["pyserial>=3.0"]

[project.optional-dependencies]
test = ["pytest>=7.0"]

[tool.setuptools.packages.find]
where = ["."]
include = ["pySerialMux*"]
